--- apps/foreldrepengeoversikt/src/types/BehandlingTilstand.ts
@@ -1,7 +1,0 @@
-export enum BehandlingTilstand {
-    TIDLIG_SØKNAD = 'VENT_TIDLIG_SØKNAD',
-    VENTER_PÅ_INNTEKTSMELDING = 'VENT_INNTEKTSMELDING',
-    VENTER_PÅ_DOKUMENTASJON = 'VENT_DOKUMENTASJON',
-    VENTER_PÅ_MELDEKORT = 'VENT_MELDEKORT',
-    UNDER_BEHANDLING = 'UNDER_BEHANDLING',
-}
