--- apps/foreldrepengeoversikt/src/types/DekningsgradDTO.ts
@@ -1,4 +1,0 @@
-export enum DekningsgradDTO {
-    ÅTTI_PROSENT = 'ÅTTI',
-    HUNDRE_PROSENT = 'HUNDRE',
-}
