--- apps/foreldrepengeoversikt/src/types/Familiehendelse.ts
@@ -1,6 +1,0 @@
-export interface Familiehendelse {
-    fødselsdato?: string;
-    termindato?: string;
-    omsorgsovertakelse?: string;
-    antallBarn: number;
-}
