--- apps/foreldrepengeoversikt/src/types/Forelder.ts
@@ -1,4 +1,0 @@
-export enum Forelder {
-    mor = 'MOR',
-    farMedmor = 'FAR_MEDMOR',
-}
