--- apps/foreldrepengeoversikt/src/types/PersonFnrDTO.ts
@@ -1,3 +1,0 @@
-export interface PersonFnrDTO {
-    fnr: string;
-}
