+++ apps/foreldrepengeoversikt/src/types/ÅpenBehandling.ts
@@ -1,19 +1,3 @@
-import { SaksperiodeNy, SøknadSVP } from '@navikt/fp-types';
-
-import { BehandlingTilstand } from './BehandlingTilstand';
-
-export interface ÅpenBehandlingFP {
-    tilstand: BehandlingTilstand;
-    søknadsperioder?: SaksperiodeNy[];
-}
-
-export interface ÅpenBehandlingES {
-    tilstand: BehandlingTilstand;
-}
-
-export interface ÅpenBehandlingSVP {
-    tilstand: BehandlingTilstand;
-    søknad: SøknadSVP;
-}
+import { ÅpenBehandlingFP, ÅpenBehandlingSVP } from '@navikt/fp-types';
 
 export type ÅpenBehandling = ÅpenBehandlingFP | ÅpenBehandlingSVP;
